Reverted build.gradle to original file and regenerated Gradle files
Replaced build.gradle with the file found on BrightSpace, and regenerated the files that Gradle would create. This fixed all of the Gradle errors we had with LychrelComputation.java

I also cast an instance of `num` to `int` in line 21 of LychrelComputation.java, since `setInputValue` only took integers, which was the only remaining error. Mentioning this in case it causes errors later.