Frontend partially implimented + database + DataStoreImpl implimented (#49)
* Removed imports for autogenerated code

If this fails the checks then I'll just delete the branch because I wasn't sure if this was needed with the code being generated by the build process (or however it is).

* Basic test of Scanner

* Started the skeleton of the system for reading files.

* Trying to fix the issue with lychrelCheck to no avail.

* Fixed DataStoreServiceImpl sort of- May need fixing

I think I implemented it wrong because it freezes when trying to send stuff to datastore.

* misc stashed changes

* Created database and fixed DataStoreServiceImpl

It now works if you input an integer! There's currently no way to view the database.